`contribute` route structure
