+++ src/routes/contribute/+page.server.ts
@@ -1,0 +1,22 @@
+import type { PageServerLoad } from './$types';
+import { prisma } from '$lib/server/prisma';
+import { error } from '@sveltejs/kit';
+
+
+export const load = (async ({ locals, params }) => {
+    const userEmail = (await locals.getSession())?.user?.email
+
+    if (!userEmail) {
+        return { userVerified: false }
+    }
+
+    const user = await prisma.user.findUnique({ where: { email: userEmail }, include: { workplaceReviewTokens: { include: { workplaceReview: { select: { id: true } } } } } })
+
+    if (!user) {
+        throw error(404, "unexpected error")
+    }
+
+    const unusedWorkplaceReviewTokens = user?.workplaceReviewTokens.filter(({ workplaceReview }) => workplaceReview === null)
+
+    return { userVerified: user.industryVerificationToken !== null, reviewToken: unusedWorkplaceReviewTokens.at(0)?.token };
+}) satisfies PageServerLoad;
