+++ src/routes/places/[placeId]/+page.svelte
@@ -1,16 +1,12 @@
 <script lang="ts">
-	import Modal from '$lib/components/Modal.svelte';
-	import ExperienceReviewForm from './ExperienceReviewForm.svelte';
-	import WorkplaceReviewForm from './WorkplaceReviewForm.svelte';
 	import Comment from './Comment.svelte';
-	import type { ActionData, PageData } from './$types';
-	import { enhance } from '$app/forms';
+	import type { PageData } from './$types';
 	import type { ReviewCommentReaction } from '@prisma/client';
 	import { slide } from 'svelte/transition';
 	import { scoreColorGradient } from '$lib/utils/colors';
+	import {page} from '$app/stores'
 
 	export let data: PageData;
-	export let form: ActionData;
 
 	let showSection: 'workplace' | 'experience' = 'workplace';
 
@@ -18,8 +14,6 @@
 		(map, cr) => map.set(cr.reviewCommentId, cr),
 		new Map<string, ReviewCommentReaction>()
 	);
-
-	const workplaceReviewToken = data.reviewToken;
 
 	const workplaceSections: { key: keyof typeof data.comments.workplace; heading: string }[] = [
 		{ key: 'general', heading: 'general comments' },
@@ -172,35 +166,10 @@
 
 	<!-- TODO: move this to the place's contribute page -->
 	{#if data.userVerified}
-		{#if data.previousWorkplaceReview}
-			<p class="text-accent">you already wrote a workplace review for this place</p>
-			<form method="post" action="?/deleteReview" use:enhance>
-				<input name="reviewId" class="hidden" value={data.previousWorkplaceReview.id} />
-				<button class="btn-outline btn-secondary btn" type="submit">delete your review</button>
-			</form>
-		{:else if workplaceReviewToken}
-			<h3>have you worked at this establishment?</h3>
-			<Modal id="workplaceReview" buttonText="workplace review">
-				<WorkplaceReviewForm
-					zodErrors={form?.workplaceReviewSchemaErrors}
-					placeId={data.place.id}
-					successfullyPosted={form?.postWorkplaceReviewSuccess}
-					{workplaceReviewToken}
-				/>
-			</Modal>
-		{:else}
-			<p>you're out of workplace reviews :(</p>
-		{/if}
 		<h3>have you visited or worked for this establishment?</h3>
-		{#if data.previousExperienceReview}
-			<p>you already wrote a visit review for this place</p>
-		{:else}
-			<Modal id="experienceReview" buttonText="write a review">
-				<ExperienceReviewForm
-					placeId={data.place.id}
-					successfullyPosted={form?.postExperienceReviewSuccess}
-				/>
-			</Modal>
-		{/if}
+
+		<p>if you would like to contribute your own experience, we would love to hear about it</p>
+
+		<a class="btn" href={`/contribute/${$page.params.placeId}`}>contribute</a>
 	{/if}
 </div>
